requirements.txt
aioredis==1.2.0
amqp==2.4.1
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
bs4==0.0.1
celery==4.2.1
certifi==2018.11.29
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
daphne==2.2.5
decorator==4.3.0
Django==2.1.7
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.1.0
djangorestframework==3.8.0
docxtpl==0.5.17
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
Jinja2==2.10
kombu==4.3.0
lxml==4.3.0
MarkupSafe==1.1.0
msgpack==0.6.1
parso==0.3.1
pbr==5.1.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
pydot==1.4.1
Pygments==2.3.1
PyHamcrest==1.9.0
pyparsing==2.3.0
python-docx==0.8.7
pytz==2018.7
redis==3.1.0
requests==2.21.0
six==1.12.0
sorl-thumbnail==12.5.0
soupsieve==1.7.2
sqlparse==0.2.4
stevedore==1.30.0
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
vine==1.2.0
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
zope.interface==4.6.0
